refactor: move anim state graph editor to its own directory
The anim state graph editor is a full-featured graph editing tool (8 files, ~1800 lines) that takes over the viewport in fullscreen mode. It was previously co-located with simple asset inspector panels in inspector/assets/. This moves the editor module to src/editor/animstategraph/ while keeping the thin inspector shell in inspector/assets/animstategraph.ts.

Made-with: Cursor
