+++ src/editor/animstategraph/anim-component.ts
