+++ src/editor/animstategraph/parameters.ts
@@ -2,8 +2,8 @@ import type { Observer } from '@playcanvas/observer';
 import { Panel, Button } from '@playcanvas/pcui';
 import { ANIM_EQUAL_TO, ANIM_PARAMETER_BOOLEAN, ANIM_PARAMETER_FLOAT, ANIM_PARAMETER_INTEGER, ANIM_PARAMETER_TRIGGER } from 'playcanvas';
 
-import type { Attribute } from '../attribute.type.d';
-import { AttributesInspector } from '../attributes-inspector';
+import type { Attribute } from '../inspector/attribute.type.d';
+import { AttributesInspector } from '../inspector/attributes-inspector';
 
 
 const CLASS_ANIMSTATEGRAPH = 'asset-animstategraph-inspector';
